:memo: Simplify docs section
Make section more concise so it's quicker to read without being any less
informative. The intro has been refined in previous commit so that it's
more descriptive (mentions testing) so we can get rid of the repeat
description here.

The note on the diagram has been moved closer to the diagram so it's
used as a description.